refactor(report): split `report` module
