+++ src/report/json.rs
@@ -1,0 +1,95 @@
+use crate::backend::trace::Trace;
+use crate::TraceInfo;
+use serde::{Serialize, Serializer};
+use trippy::dns::{DnsResolver, Resolver};
+
+/// Generate a CSV report of trace data.
+pub fn report(
+    info: &TraceInfo,
+    report_cycles: usize,
+    resolver: &DnsResolver,
+) -> anyhow::Result<()> {
+    let trace = super::wait_for_round(&info.data, report_cycles)?;
+    let hops: Vec<ReportHop> = trace
+        .hops(Trace::default_flow_id())
+        .iter()
+        .map(|hop| {
+            let hosts: Vec<_> = hop
+                .addrs()
+                .map(|ip| Host {
+                    ip: ip.to_string(),
+                    hostname: resolver.reverse_lookup(*ip).to_string(),
+                })
+                .collect();
+            ReportHop {
+                ttl: hop.ttl(),
+                hosts,
+                loss_pct: hop.loss_pct(),
+                sent: hop.total_sent(),
+                last: hop.last_ms().unwrap_or_default(),
+                recv: hop.total_recv(),
+                avg: hop.avg_ms(),
+                best: hop.best_ms().unwrap_or_default(),
+                worst: hop.worst_ms().unwrap_or_default(),
+                stddev: hop.stddev_ms(),
+            }
+        })
+        .collect();
+
+    let report = Report {
+        info: ReportInfo {
+            target: Host {
+                ip: info.target_addr.to_string(),
+                hostname: info.target_hostname.to_string(),
+            },
+        },
+        hops,
+    };
+    println!("{}", serde_json::to_string_pretty(&report).unwrap());
+    Ok(())
+}
+
+#[derive(Serialize)]
+pub struct Report {
+    pub info: ReportInfo,
+    pub hops: Vec<ReportHop>,
+}
+
+#[derive(Serialize)]
+pub struct ReportInfo {
+    pub target: Host,
+}
+
+#[derive(Serialize)]
+pub struct ReportHop {
+    ttl: u8,
+    hosts: Vec<Host>,
+    #[serde(serialize_with = "fixed_width")]
+    loss_pct: f64,
+    sent: usize,
+    #[serde(serialize_with = "fixed_width")]
+    last: f64,
+    recv: usize,
+    #[serde(serialize_with = "fixed_width")]
+    avg: f64,
+    #[serde(serialize_with = "fixed_width")]
+    best: f64,
+    #[serde(serialize_with = "fixed_width")]
+    worst: f64,
+    #[serde(serialize_with = "fixed_width")]
+    stddev: f64,
+}
+
+#[derive(Serialize)]
+pub struct Host {
+    pub ip: String,
+    pub hostname: String,
+}
+
+#[allow(clippy::trivially_copy_pass_by_ref)]
+fn fixed_width<S>(val: &f64, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&format!("{val:.2}"))
+}
